Multiple updates to lookup modules.
Made changes to the config and DB schema to support dynamic lookup
sources. Modified the lookup modules to support the DB schema
changes.

Setting the modules up to allow a lookup service to run each module
by calling the same method. This will allow additional modules to
be easily added.